+++ packages/perseus/src/widgets/__stories__/interactive-graph.stories.tsx
@@ -18,7 +18,7 @@ import {
     segmentWithAllLockedLineVariations,
     segmentWithAllLockedRayVariations,
     sinusoidQuestion,
-    segmentWithLockedCircles,
+    segmentWithLockedEllipses,
     segmentWithLockedVectors,
 } from "../__testdata__/interactive-graph.testdata";
 
@@ -117,8 +117,11 @@ export const AllLockedRays = (args: StoryArgs): React.ReactElement => (
     />
 );
 
-export const LockedCircle = (args: StoryArgs): React.ReactElement => (
-    <RendererWithDebugUI {...mafsOptions} question={segmentWithLockedCircles} />
+export const LockedEllipse = (args: StoryArgs): React.ReactElement => (
+    <RendererWithDebugUI
+        {...mafsOptions}
+        question={segmentWithLockedEllipses}
+    />
 );
 
 export const LockedVector = (args: StoryArgs): React.ReactElement => (