+++ packages/perseus/src/widgets/__testdata__/interactive-graph.testdata.ts
@@ -1975,18 +1975,18 @@ export const segmentWithLockedFigures: PerseusRenderer =
     interactiveGraphQuestionBuilder()
         .addLockedPointAt(-7, -7)
         .addLockedLine([-7, -5], [2, -3])
-        .addLockedCircle([0, 5], 2)
+        .addLockedEllipse([0, 5], 2)
         .build();
 
-export const segmentWithLockedCircles: PerseusRenderer =
+export const segmentWithLockedEllipses: PerseusRenderer =
     interactiveGraphQuestionBuilder()
-        .addLockedCircle([0, 0], 5)
-        .addLockedCircle([-5, 5], 2, {
+        .addLockedEllipse([0, 0], 5)
+        .addLockedEllipse([-5, 5], 2, {
             color: "green",
             fillStyle: "solid",
             strokeStyle: "solid",
         })
-        .addLockedCircle([5, 5], 2, {
+        .addLockedEllipse([5, 5], 2, {
             color: "green",
             fillStyle: "translucent",
             strokeStyle: "dashed",